requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
Django==2.2.4
django-cors-headers==3.0.2
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.10.2
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
mercadopago==0.3.5
openapi-codec==1.3.2
Pillow==6.1.0
psycopg2-binary==2.8.3
pycparser==2.19
pyOpenSSL==19.0.0
pytz==2019.2
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
stripe==2.35.0
uritemplate==3.0.0
urllib3==1.25.3
